requirements.txt
cryptography>=42.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
pytest>=8.0.0
